refactor(db): move repositories to repositories/db, userroles logic to services/db
Addresses review on #832:
- controllers/db/*.repository.ts -> repositories/db/*.repository.ts
- userroles.controller.ts had no req/res handler, only validation and
  allow-list mapping -> services/db/userroles.service.ts

Imports updated in the two remaining controllers, samlUtils.js and
[...nextauth].jsx. No behavior or query changes.
